basm: huge refactoring to allow float values

+++ cpclib-asm/src/parser/obtained.rs
@@ -264,7 +264,7 @@ impl LocatedToken {
 
 
                         if offset.is_some() {
-                            let offset = offset.as_ref().unwrap().eval()? as usize;
+                            let offset = offset.as_ref().unwrap().eval()?.int() as usize;
                             if offset >= data.len() {
                                 return Err(AssemblerError::AssemblingError {
                                     msg: format!("Unable to read {:?}. Only {} are available", fname, data.len())
@@ -275,7 +275,7 @@ impl LocatedToken {
 
 
                         if length.is_some() {
-                            let length = length.as_ref().unwrap().eval()? as usize;
+                            let length = length.as_ref().unwrap().eval()?.int() as usize;
                             data = &data[..length];
                             if data.len() != length {
                                 return Err(AssemblerError::AssemblingError {
